# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MiniStay API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reviews/{room_id}": {
            "get": {
                "tags": [
                    "User - Reviews"
                ],
                "summary": "Get visible reviews for a room",
                "description": "Return only reviews where is_visible = true",
                "operationId": "dc746d20fa2ee4c4d80718388ec5a06d",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "path",
                        "description": "ID of the room",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of reviews",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Review"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/admin/reviews/{id}/moderate": {
            "patch": {
                "tags": [
                    "Admin - Reviews"
                ],
                "summary": "Update review visibility",
                "description": "Admin can hide or show user reviews",
                "operationId": "41fab898634fb290e03387500d80dee0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Review ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "is_visible": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Visibility updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Review visibility updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "adminAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Review": {
                "required": [
                    "booking_id",
                    "user_id",
                    "room_id",
                    "rating"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "booking_id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "room_id": {
                        "type": "integer"
                    },
                    "rating": {
                        "type": "integer"
                    },
                    "comment": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_visible": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}